CancelObject: Grow object_id type in Connect
We aim to increase maximum cancellable object count to thousands, so uint8_t won't do.

BFW-6465

+++ src/connect/command.cpp
@@ -400,8 +400,8 @@ Command Command::parse_json_command(CommandId id, char *body, size_t body_size, 
         } else if (is_arg("dialog_id", Type::Primitive)) {
             INT_ARG(DialogAction, uint32_t, dialog_id, ArgDialogId)
         } else if (is_arg("id", Type::Primitive)) {
-            INT_ARG(CancelObject, uint8_t, id, ArgId)
-            INT_ARG(UncancelObject, uint8_t, id, ArgId)
+            INT_ARG(CancelObject, uint16_t, id, ArgId)
+            INT_ARG(UncancelObject, uint16_t, id, ArgId)
 #if PRINTER_IS_PRUSA_COREONE() || defined(UNITTESTS)
         } else if (is_arg("chamber.target_temp", Type::Primitive)) {
             SET_VALUE_ARG(PropertyName::ChamberTargetTemp, uint32_t);
