+++ src/connect/error_printer.hpp
@@ -38,7 +38,7 @@ private:
     virtual void init_connect(const char *token) override;
     virtual uint32_t cancelable_fingerprint() const override;
 #if HAS_CANCEL_OBJECT()
-    virtual void set_object_cancelled(uint8_t, bool) override;
+    virtual void set_object_cancelled(uint16_t, bool) override;
 #endif
     virtual void reset_printer() override;
     virtual const char *dialog_action(printer_state::DialogId dialog_id, Response response) override;
