+++ tests/unit/connect/mock_printer.h
@@ -138,7 +138,7 @@ public:
     virtual void set_slot_info(size_t, const SlotInfo &) override {}
 
 #if HAS_CANCEL_OBJECT()
-    virtual void set_object_cancelled(uint8_t, bool) override {}
+    virtual void set_object_cancelled(uint16_t, bool) override {}
 #endif
 };
